src/requirements.txt
click==8.1.7
dataclasses-json==0.5.7
elicited==0.0.5
libcst==1.5.1
marshmallow==3.23.1
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
numpy==2.0.2
packaging==24.2
psutil==6.1.0
Pygments==2.18.0
pyre-extensions==0.0.32
python-frontmatter==1.1.0
PyYAML==6.0.2
scipy==1.13.1
tabulate==0.9.0
TestSlide==2.7.1
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.12.2
